[refiners][persona] - renamed SpService to UserService
