+++ search-parts/src/services/UserService/MockUserService.ts
@@ -1,44 +1,31 @@
 // PnP
 import { ConsoleListener, LogLevel, Logger } from '@pnp/logging';
-import { sp, SPRest } from '@pnp/sp';
-
-// SPFx
-import { PageContext } from '@microsoft/sp-page-context';
 
 // Interface
-import { IUser, IUserProfileProperty } from './../../models/IUser';
+import { IUser, IUserProfileProperty } from '../../models/IUser';
 import IUserService from './IUserService';
 
+// MockData
+import { mockUserData } from './MockUserData';
+
 // Class
-export class UserService implements IUserService {
-
-  private _pageContext: PageContext;
-  private _localPnPSetup: SPRest;
-
-  constructor(pageContext: PageContext, ) {
-    this._pageContext = pageContext;
-
+export class MockUserService implements IUserService {
+  constructor() {
     const consoleListener = new ConsoleListener();
     Logger.subscribe(consoleListener);
-
-    this._localPnPSetup = sp.configure({
-      headers: {
-        Accept: 'application/json; odata=nometadata',
-      },
-    }, this._pageContext.web.absoluteUrl);
   }
 
   /**
    * Get user information
    */
   public async getUserProperties(accountName: string): Promise<IUser> {
-    let response : any;
+    let response: any;
     try {
-      response = await this._localPnPSetup.profiles.getPropertiesFor(accountName);
+      response = await mockUserData;
       if (!!response && !response['odata.null']) {
         let properties = {};
         if (!!response.UserProfileProperties) {
-          response.UserProfileProperties.forEach((prop : IUserProfileProperty) => {
+          response.UserProfileProperties.forEach((prop: IUserProfileProperty) => {
             properties[prop.Key] = prop.Value;
           });
           response.userProperties = properties;
